requirements.txt
quart==0.19.9 #0.19.4 flask==3.0.3
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
supabase==2.16.0
httpx==0.27.2
hypercorn==0.15.0
ollama==0.3.3

# Optional but recommended
aiohttp==3.9.1
python-multipart==0.0.6

